third_party/WebKit/LayoutTests/fast/filesystem/null-arguments-expected.txt

This test tries calling various filesystem functions with null arguments.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Successfully obtained FileSystem: file__0:Temporary
PASS writer.error is null
PASS writer.write(null) threw exception TypeError: Failed to execute 'write' on 'FileWriter': parameter 1 is not of type 'Blob'..
PASS writer.error is null
PASS successfullyParsed is true

TEST COMPLETE
